feat(gtfs-parser): rewrite parser to pluggable LMDB dedup
Introduce LMDB-backed state management and a new data model for
per-entity updates
(LmdbEntry, PendingUpdate, ProcessingAccumulator) and enhanced repo
(GtfsStateRepository) with thread-local buffers

Add pluggable, field-based dispatch for GTFS FeedEntity payloads
(FeedEntityHandler, TripUpdateHandler, HashBasedTripUpdateFilter) and
new GtfsEntityType and GtfsByteFeedWrapper

Implement a new GTFS parser orchestration
(GtfsParserService) with read/write phases and a dispatch table
plus detailed Micrometer metrics

Rename and augment utilities
(FeedHashing, ProtoUtils, SizeFormat changes) to support new flow

Change LMDB storage path to ~/.cache/lmdb and drop "./lmdb-data"
MDB_NOSYNC optimization

Remove legacy producer path in favor of LMDB-based dedup
(GtfsProducerOrchestrator removed)

Note: downstream code must adapt to new parse output wrapper and API

BREAKING CHANGE: rework GTFS parser API and remove the old
producer orchestrator. The parser now returns a GtfsByteFeedWrapper,
and uses a pluggable, LMDB-backed dedup flow. Downtime clients
must adapt.

+++ src/main/java/org/example/sirianalyzer/util/SizeFormat.java
@@ -21,7 +21,7 @@ public final class SizeFormat {
      * @return A human-readable string representation of the input bytes using
      *         binary prefixes
      */
-    public static String humanBytes(int bytes) {
+    public static String humanBytes(long bytes) {
         if (bytes < 0)
             bytes = 0;
         if (bytes < 1024)
